refactor(state_detect): consolidate state-detection helpers into leaf module
Move 18 screenshot->bool detectors out of nav.py, fight.py,
card_mastery_state.py, and deck_utils.py into a new pyclashbot/bot/state_detect.py
leaf module. All callers across pyclashbot/ and tests/ import directly from
state_detect; nav.py keeps re-exports for transitional compatibility.
