+++ pyclashbot/bot/deck_randomization.py
@@ -4,16 +4,16 @@ Deck randomization methods for py-clash-bot.
 
 import time
 
-from pyclashbot.bot.deck_utils import (
-    is_single_deck_layout_by_pixel,
-    randomize_and_check_deck,
-)
+from pyclashbot.bot.deck_utils import randomize_and_check_deck
 from pyclashbot.bot.nav import (
     DECK_TABS_REGION,
-    check_if_on_clash_main_menu,
     get_to_card_page_from_clash_main,
     return_to_clash_main_from_card_page,
     switch_deck_page,
+)
+from pyclashbot.bot.state_detect import (
+    check_if_on_clash_main_menu,
+    is_single_deck_layout_by_pixel,
 )
 from pyclashbot.detection.image_rec import find_image
 from pyclashbot.utils.cancellation import interruptible_sleep
