+++ pyclashbot/bot/nav.py
@@ -3,13 +3,39 @@ import time
 from typing import Literal
 
 from pyclashbot.bot.constants import CLASH_MAIN_DEADSPACE_COORD as CLASH_MAIN_MENU_DEADSPACE_COORD
+from pyclashbot.bot.state_detect import (
+    check_for_post_battle_button,
+    check_for_trophy_reward_menu,
+    check_if_battle_mode_is_selected,
+    check_if_in_battle,
+    check_if_on_battle_log_page,
+    check_if_on_card_page,
+    check_if_on_clan_chat,
+    check_if_on_clash_main_burger_button_options_menu,
+    check_if_on_clash_main_menu,
+    check_if_on_shop,
+    check_if_on_social,
+)
 from pyclashbot.detection.image_rec import (
-    all_pixels_are_equal,
     find_image,
     pixel_is_equal,
 )
 from pyclashbot.utils.cancellation import interruptible_sleep
 from pyclashbot.utils.logger import Logger
+
+# Re-export moved detectors so existing `from pyclashbot.bot.nav import check_if_*`
+# call sites keep working. New code should import from pyclashbot.bot.state_detect.
+__all__ = [
+    "check_if_battle_mode_is_selected",
+    "check_if_in_battle",
+    "check_if_on_battle_log_page",
+    "check_if_on_card_page",
+    "check_if_on_clan_chat",
+    "check_if_on_clash_main_burger_button_options_menu",
+    "check_if_on_clash_main_menu",
+    "check_if_on_shop",
+    "check_if_on_social",
+]
 
 CLASH_MAIN_OPTIONS_BURGER_BUTTON = (390, 62)
 BATTLE_LOG_BUTTON = (241, 43)
@@ -78,122 +104,6 @@ def check_for_in_battle_with_delay(emulator) -> bool:
     return False
 
 
-def check_if_in_battle(emulator):
-    iar_bgr = emulator.screenshot()
-    if iar_bgr is None:
-        return False
-
-    # Convert to RGB for easier reasoning about expected colors.
-    iar = iar_bgr[..., ::-1]
-
-    def get_pixel(y: int, x: int) -> list[int] | None:
-        if y >= iar.shape[0] or x >= iar.shape[1]:
-            return None
-        return iar[y][x].tolist()
-
-    def is_bright(pixel: list[int] | None, threshold: int = 180) -> bool:
-        if pixel is None:
-            return False
-        return all(channel >= threshold for channel in pixel) or (
-            150 <= pixel[0] <= 195 and 40 <= pixel[1] <= 60 and 40 <= pixel[2] <= 60
-        )
-
-    def is_filled_crown(pixel: list[int] | None) -> bool:
-        """Check if pixel is a gold/yellow filled crown or UI accent."""
-        if pixel is None:
-            return False
-        r, g, b = pixel
-        return r >= 170 and g >= 130 and b <= 140
-
-    def is_scoreboard_purple(pixel: list[int] | None) -> bool:
-        if pixel is None:
-            return False
-        r, g, b = pixel
-        return r >= 200 and b >= 200 and g <= 140
-
-    def check_mode(coords: list[tuple[int, int]]) -> bool:
-        pixels = [get_pixel(y, x) for y, x in coords]
-        # Allow one of the "bright" UI pixels to change (crowns filling, overlays,
-        # small rendering differences) while still requiring the purple scoreboard.
-        bright_required = max(1, len(coords) - 2)
-        bright_count = sum(1 for pixel in pixels[:-1] if is_bright(pixel) or is_filled_crown(pixel))
-        return bright_count >= bright_required and is_scoreboard_purple(pixels[-1])
-
-    # When the emote is closed these pixels are not considered bright:
-    # (533, 80) RGB=[157, 44, 44] and (532, 77) RGB=[187, 55, 55]
-    coords_1v1 = [(528, 49), (532, 77), (546, 52), (546, 77), (618, 115)]
-    coords_2v2 = [(534, 53), (533, 80), (548, 52), (548, 76), (615, 114)]
-
-    if check_mode(coords_1v1):
-        return True
-    if check_mode(coords_2v2):
-        return True
-
-    return False
-
-
-def check_for_post_battle_button(emulator) -> bool:
-    """Checks for the post-battle OK/Exit buttons (battle-end screen)."""
-    image = emulator.screenshot()
-    if image is None:
-        return False
-
-    if find_image(image, "ok_post_battle_button", tolerance=0.85) is not None:
-        return True
-
-    if find_image(image, "exit_battle_button", tolerance=0.9) is not None:
-        return True
-
-    return False
-
-
-def check_if_battle_has_ended(emulator) -> bool:
-    """Best-effort confirmation that a battle ended (avoid false positives mid-fight)."""
-    if check_if_on_clash_main_menu(emulator):
-        return True
-
-    if check_for_trophy_reward_menu(emulator):
-        return True
-
-    if check_for_post_battle_button(emulator):
-        return True
-
-    return False
-
-
-def check_for_trophy_reward_menu(emulator) -> bool:
-    iar = emulator.screenshot()
-
-    pixels = [
-        iar[592][172],
-        iar[617][180],
-        iar[607][190],
-        iar[603][200],
-        iar[596][210],
-        iar[593][220],
-        iar[600][230],
-        iar[610][235],
-        iar[623][246],
-    ]
-    colors = [
-        [255, 184, 68],
-        [255, 175, 78],
-        [255, 175, 78],
-        [248, 239, 227],
-        [255, 187, 104],
-        [255, 176, 79],
-        [255, 187, 104],
-        [255, 175, 78],
-        [253, 135, 39],
-    ]
-
-    for i, pixel in enumerate(pixels):
-        if not pixel_is_equal(pixel, colors[i], tol=25):
-            return False
-
-    return True
-
-
 def handle_trophy_reward_menu(
     emulator,
     logger: Logger,
@@ -246,78 +156,6 @@ def wait_for_clash_main_menu(emulator, logger: Logger, deadspace_click=True) -> 
     return True
 
 
-def check_if_on_clash_main_menu(emulator) -> bool:
-    """Checks if the user is on the clash main menu.
-    Returns True if on main menu, False if not.
-    """
-    image = emulator.screenshot()
-    pixels = [
-        image[14][209],  # white
-        image[14][325],  # white
-        image[19][298],  # yellow
-        image[17][399],  # green
-        image[581][261],  # green
-        image[584][166],  # bluegrey
-        image[621][166],  # bluegrey
-    ]
-
-    # google play colors
-    colors_1 = [
-        [255, 255, 255],
-        [255, 255, 255],
-        [53, 199, 233],
-        [25, 198, 65],
-        [138, 105, 71],
-        [139, 105, 72],
-        [155, 120, 82],
-    ]
-
-    # memu colors
-    colors_2 = [
-        [255, 255, 255],
-        [255, 255, 255],
-        [53, 200, 233],
-        [24, 199, 65],
-        [138, 105, 71],
-        [139, 105, 72],
-        [155, 120, 81],
-    ]
-
-    # post Clash Royale app update colors
-    colors_3 = [
-        [57, 151, 206],
-        [21, 148, 42],
-        [36, 17, 1],
-        [231, 190, 123],
-        [140, 105, 74],
-        [140, 105, 74],
-        [156, 121, 82],
-    ]
-
-    # print("{:^15} | {:^15} | {:^15}".format("Seen", "Google", "Memu"))
-    # for seen_pixel, google_play_color, memu_color in zip(pixels, colors_1, colors_2):
-    #     seen_pixel =str(seen_pixel[0])+ ' '+ str(seen_pixel[1])+ ' '+ str(seen_pixel[2])
-    #     google_play_color = (
-    #         str(google_play_color[0]) + ' ' +
-    #         str(google_play_color[1]) + ' ' +
-    #         str(google_play_color[2])
-    #     )
-    #     memu_color = str(memu_color[0]) + ' ' + str(memu_color[1]) + ' ' + str(memu_color[2])
-    #     print(
-    #         "{:^15} | {:^15} | {:^15}".format(seen_pixel, google_play_color, memu_color)
-    #     )
-
-    for colors in [colors_1, colors_2, colors_3]:
-        if all_pixels_are_equal(
-            pixels,
-            colors,
-            25,
-        ):
-            return True
-
-    return False
-
-
 def get_to_card_page_from_clash_main(
     emulator,
     logger: Logger,
@@ -348,62 +186,6 @@ def get_to_card_page_from_clash_main(
     logger.change_status(status="Made it to card page")
 
     return "good"
-
-
-def check_if_on_card_page(emulator) -> bool:
-    iar = emulator.screenshot()
-
-    pixels = [
-        iar[433][58],
-        iar[116][59],
-        iar[58][82],
-        iar[64][179],
-        iar[62][108],
-        iar[67][146],
-        iar[77][185],
-        iar[77][84],
-    ]
-
-    colors1 = [
-        [222, 0, 235],
-        [255, 255, 255],
-        [203, 137, 44],
-        [195, 126, 34],
-        [255, 255, 255],
-        [255, 255, 255],
-        [177, 103, 15],
-        [178, 104, 15],
-    ]
-
-    colors2 = [
-        [220, 0, 234],
-        [255, 255, 255],
-        [209, 68, 41],
-        [202, 64, 41],
-        [255, 255, 255],
-        [255, 255, 255],
-        [185, 52, 41],
-        [185, 52, 41],
-    ]
-
-    def pixel_to_string(pixel):
-        return f"[{pixel[0]},{pixel[1]},{pixel[2]}],"
-
-    # print("{:^17} {:^17} {:^17}".format("pixel", "color1", "color2"))
-    # for pixel, color1, color2 in zip(pixels, colors1, colors2):
-    #     print(
-    #         "{:^17} {:^17} {:^17}".format(
-    #             pixel_to_string(pixel), pixel_to_string(color1), pixel_to_string(color2)
-    #         )
-    #     )
-
-    if all_pixels_are_equal(pixels, colors1, tol=25):
-        return True
-
-    if all_pixels_are_equal(pixels, colors2, tol=25):
-        return True
-
-    return False
 
 
 def return_to_clash_main_from_card_page(emulator, logger: Logger) -> bool:
@@ -490,70 +272,6 @@ def wait_for_battle_log_page(
     return "good"
 
 
-def check_if_on_battle_log_page(emulator) -> bool:
-    iar = emulator.screenshot()
-
-    pixels = [
-        iar[72][160],
-        iar[71][187],
-        iar[71][197],
-        iar[72][231],
-        iar[73][258],
-        iar[64][366],
-        iar[79][365],
-        iar[70][365],
-        iar[62][92],
-        iar[77][316],
-    ]
-    colors = [
-        [255, 255, 255],
-        [255, 255, 255],
-        [255, 255, 255],
-        [255, 255, 255],
-        [255, 255, 255],
-        [147, 135, 254],
-        [38, 38, 240],
-        [255, 255, 255],
-        [138, 122, 115],
-        [124, 106, 99],
-    ]
-
-    for i, p in enumerate(pixels):
-        if not pixel_is_equal(p, colors[i], tol=25):
-            return False
-    return True
-
-
-def check_if_on_clash_main_burger_button_options_menu(emulator) -> bool:
-    iar = emulator.screenshot()
-    pixels = [
-        iar[42][256],
-        iar[41][275],
-        iar[41][282],
-        iar[42][293],
-        iar[44][325],
-        iar[32][239],
-        iar[34][336],
-        iar[50][248],
-        iar[49][336],
-    ]
-    colors = [
-        [255, 255, 255],
-        [255, 255, 255],
-        [255, 255, 255],
-        [255, 255, 254],
-        [255, 255, 255],
-        [255, 187, 105],
-        [255, 187, 105],
-        [255, 175, 78],
-        [255, 175, 78],
-    ]
-    for i, color in enumerate(colors):
-        if not pixel_is_equal(pixels[i], color, tol=25):
-            return False
-    return True
-
-
 def wait_for_clash_main_burger_button_options_menu(
     emulator,
     logger: Logger,
@@ -595,48 +313,6 @@ def wait_for_clash_main_burger_button_options_menu(
     return "good"
 
 
-def check_if_battle_mode_is_selected(emulator, mode: str):
-    """Checks if the given battle mode is selected on the clash main menu.
-
-    Args:
-        emulator: The emulator controller.
-        mode: The battle mode to check for.
-
-    Returns:
-        True if the mode is selected, False otherwise.
-    """
-    expected_mode_types = ["Classic 1v1", "Classic 2v2", "Trophy Road"]
-
-    # Check if the mode is valid
-    if mode not in expected_mode_types:
-        print(f'[!] Fatal error: Mode "{mode}" is not a valid mode type. Expected one of {expected_mode_types}.')
-        return None
-
-    mode2folder = {
-        "Classic 1v1": "selected_1v1_on_main",
-        "Classic 2v2": "selected_2v2_on_main",
-        "Trophy Road": "selected_trophy_road_on_main",
-    }
-
-    look_folder = mode2folder[mode]
-
-    print(f"[DEBUG] Checking if {mode} is selected...")
-    print(f"[DEBUG] Looking in folder: {look_folder}")
-    print("[DEBUG] Subcrop: (270, 455, 350, 533)")
-
-    # find image on screen
-    coord = find_image(
-        emulator.screenshot(),
-        look_folder,
-        tolerance=0.9,
-        subcrop=(270, 455, 350, 533),
-    )
-
-    print(f"[DEBUG] Found at: {coord}")
-
-    return coord is not None
-
-
 def find_fight_mode_icon(emulator, mode: str):
     expected_mode_types = ["Classic 1v1", "Classic 2v2", "Trophy Road"]
 
@@ -654,11 +330,6 @@ def find_fight_mode_icon(emulator, mode: str):
     look_folder = mode2folder[mode]
 
     image = emulator.screenshot()
-
-    # os.makedirs('select_mode_images', exist_ok=True)
-    # file_name = f'{random.randint(0,100000)}.png'
-    # file_path = os.path.join('select_mode_images', file_name)
-    # cv2.imwrite(file_path, image)
 
     fight_mode_1v1_button_location = find_image(
         image,
@@ -740,74 +411,6 @@ PAGE_CARD = "card_page"
 PAGE_SHOP = "shop"
 PAGE_SOCIAL = "social"
 PAGE_CLAN_CHAT = "clan-chat"
-
-
-def check_if_on_shop(emulator) -> bool:
-    iar = emulator.screenshot()
-    pixels = [
-        iar[589][16],
-        iar[609][19],
-        iar[14][212],
-        iar[15][331],
-        iar[600][128],
-    ]
-    colors = [
-        [140, 107, 73],
-        [151, 116, 77],
-        [57, 160, 214],
-        [22, 186, 59],
-        [255, 225, 137],
-    ]
-    for i, p in enumerate(pixels):
-        if not pixel_is_equal(p, colors[i], tol=25):
-            return False
-    return True
-
-
-def check_if_on_social(emulator) -> bool:
-    iar = emulator.screenshot()
-    pixels = [
-        iar[601][218],
-        iar[577][342],
-        iar[620][299],
-    ]
-    colors = [
-        [255, 226, 138],
-        [142, 108, 75],
-        [155, 120, 82],
-    ]
-    for i, p in enumerate(pixels):
-        if not pixel_is_equal(p, colors[i], tol=25):
-            return False
-    return True
-
-
-def check_if_on_clan_chat(emulator) -> bool:
-    iar = emulator.screenshot()
-    pixels = [
-        iar[5][5],
-        iar[628][415],
-        iar[612][177],
-        iar[44][322],
-        iar[19][321],
-        iar[620][254],
-        iar[612][161],
-        iar[584][204],
-    ]
-    colors = [
-        [141, 85, 69],
-        [105, 85, 71],
-        [255, 175, 78],
-        [237, 215, 201],
-        [245, 231, 222],
-        [107, 86, 72],
-        [109, 87, 73],
-        [160, 128, 108],
-    ]
-    for i, p in enumerate(pixels):
-        if not pixel_is_equal(p, colors[i], tol=25):
-            return False
-    return True
 
 
 MAIN_PAGE_CHECKS = {
@@ -984,30 +587,4 @@ def get_to_main_after_fight(emulator, logger):
 
 
 if __name__ == "__main__":
-    # from pyclashbot.emulators.memu import MemuEmulatorController
-    # from pyclashbot.utils.logger import Logger
-    # import cv2
-    # import os
-
-    # print("Creating logger...")
-    # logger = Logger()
-
-    # print("Creating MEmu emulator controller in DEBUG mode (no restart)...")
-    # emulator = MemuEmulatorController(logger, render_mode="directx", debug_mode=True)
-
-    # # Save a screenshot for debugging
-    # print("\nSaving screenshot for debugging...")
-    # os.makedirs("debug_screenshots", exist_ok=True)
-    # screenshot = emulator.screenshot()
-    # cv2.imwrite("debug_screenshots/current_screen.png", screenshot)
-    # print(f"Screenshot saved to: debug_screenshots/current_screen.png")
-    # print(f"Screenshot size: {screenshot.shape}")
-
-    # print("\nTesting find_fight_mode_icon...")
-    # x = check_if_battle_mode_is_selected(emulator, "Classic 2v2")
-    # print(f"Result: Is Classic 2v2 selected? {x}")
-
-    # select_mode(emulator, "Classic 1v1")
-
-    # here matt if you want to test this and see debug info
     pass
