+++ pyclashbot/emulators/bluestacks.py
@@ -8,7 +8,7 @@ import time
 from contextlib import suppress
 from os.path import normpath
 
-from pyclashbot.bot.nav import check_if_on_clash_main_menu
+from pyclashbot.bot.state_detect import check_if_on_clash_main_menu
 from pyclashbot.emulators.adb_base import AdbBasedController
 from pyclashbot.utils.cancellation import interruptible_sleep
 from pyclashbot.utils.platform import Platform, is_macos
